% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNicheOverlapProbability.R
\name{calculateNicheOverlapProbability}
\alias{calculateNicheOverlapProbability}
\title{Generate subsample Moran's I values for niche overlap given two species occurrence data.}
\usage{
calculateNicheOverlapProbability(species_models, background_data,
  predictors, iterations)
}
\arguments{
\item{species_models}{A list of species lists where each contains a path, predictive_map, occurrence_data, and args element.}

\item{background_data}{Dataframe of selected background coordinates and extracted environmental predictor data.}

\item{predictors}{A RasterStack of RasterLayer objects representing the environmental predictors.}

\item{iterations}{Number of random replicates to run for each species combination.}
}
\value{
A dataframe containing the species combination, the real Moran's I, all replicate Moran's I values, and a P value.
}
\description{
Generate subsample Moran's I values for niche overlap given two species occurrence data.
}
