% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBiasDensity.R
\name{createBiasDensity}
\alias{createBiasDensity}
\title{Create a probability density map based on the occurence points of all species.}
\usage{
createBiasDensity(predictors, occurrence_points_list, scale, shape,
  output_folder)
}
\arguments{
\item{predictors}{A RasterStack of RasterLayer objects representing the environmental predictors.}

\item{occurrence_points_list}{A list of occurence point dataframes for each species sampled.}

\item{scale}{A 2 element vector containing the minimum and maximum range for the bias probablity scale.}

\item{shape}{A shapefile to overlay on the plotted bias density map.}

\item{output_folder}{Location to save data and plots.}
}
\value{
A RasterLayer of selection probablities ranging from 0 to 100 throughout the training extent.
}
\description{
The creation of a probablity density map (bias file) helps avoid sampling bias in models. Restricting background point selection
to areas around occurence points decreases the potential of unbalanced sampling between occurence points and background points. This
may occur if not all areas in the training extent were uniformly randomly sampled.
}
