% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOccurrencePoints.R
\name{loadOccurrencePoints}
\alias{loadOccurrencePoints}
\title{Loads the occurrence data from the specified folder.}
\usage{
loadOccurrencePoints(occurrence_folder, predictors, categoricals)
}
\arguments{
\item{occurrence_folder}{A path to a directory that contains the occurrence files.}

\item{predictors}{A RasterStack of RasterLayer objects representing the environmental predictors.}

\item{categoricals}{A vector of indicies specifying the environmental predictors that contain catagorical or discrete data.}
}
\value{
List of dataframes of occurrence coordinates and extracted environmental predictor data.
}
\description{
Occurrence coordinates should be in a CSV format with the file extension '.csv'. Files should contain two columns,
longitude and latitude. Coordinates will be paired with environmental data for each environmental predictor provided.
}
