% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleNicheOverlap.R
\name{subsampleNicheOverlap}
\alias{subsampleNicheOverlap}
\title{Generate subsample values for niche overlap given two species occurrence data.}
\usage{
subsampleNicheOverlap(occurrence_data1, args1, occurrence_data2, args2,
  background_data, predictors, iterations)
}
\arguments{
\item{occurrence_data1}{An occurrence dataframe containing cooridnates and environmental data at each coordinate.}

\item{args1}{Arguments, including hyperparameters, to use when training the first pseudo model.}

\item{occurrence_data2}{An occurrence dataframe containing cooridnates and environmental data at each coordinate.}

\item{args2}{Arguments, including hyperparameters, to use when training the second pseudo model.}

\item{background_data}{Dataframe of selected background coordinates and extracted environmental predictor data.}

\item{predictors}{A RasterStack of RasterLayer objects representing the environmental predictors.}

\item{iterations}{Number of subsample values to generate.}
}
\value{
A list of subsampled niche overlap values.
}
\description{
Generate subsample values for niche overlap given two species occurrence data.
}
