% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneRegularization.R
\name{tuneRegularization}
\alias{tuneRegularization}
\title{Use the AdaLipo technique to search for the best value for the regularization hyperparameter.}
\usage{
tuneRegularization(occurrence_data, background_data, predictors,
  group_data, factor, range, num_iterations)
}
\arguments{
\item{occurrence_data}{Dataframe of species occurrence coordinates and extracted environmental predictor data.}

\item{background_data}{Dataframe of selected background coordinates and extracted environmental predictor data.}

\item{predictors}{A RasterStack of RasterLayer objects representing the environmental predictors.}

\item{group_data}{A named list of bin designation for occurrence points and background points.}

\item{factor}{The factor hyperparameter value to use in model creation.}

\item{range}{A 2 element vector containing the start and end point of the range of regularization values to search.}

\item{num_iterations}{Number of models to create.}
}
\value{
A list of results from all created models.
}
\description{
Use the AdaLipo technique to search for the best value for the regularization hyperparameter.
}
